refactor(filter): extract the no-transform detection into one authoritative header (#270)
The Cache-Control no-transform detection -- the quote-aware
directive-segment walk, the '=' directive-name cut, and the
whole-headers_out-list check -- moves verbatim from the filter module
into src/ngx_http_zstd_cache_control.h, the ngx_http_zstd_frame_probe.h
pattern applied to the next-smallest shared surface.

Why this family next: it is the most recently drift-prone one. #251
(the detection itself) and then #274 (the quoted-comma segment fix)
were each hand-mirrored into the compression branch and the brotli
hardened fork within days of merging here -- three parallel patches
per fix, and lizard counts the pair at CCN 15+8 in this repo against
an already-diverged CCN 22+9 in the compression branch's transplant
(equivalent behaviour, different shapes: this repo's quote-awareness
lives in the directive_end helper, the transplant's inline). A
consumer that includes the header instead of carrying a copy takes
the next fix at sync time with no port.

The three function bodies are byte-identical to the ones deleted from
the filter module (verified mechanically); the one delta is `static`
-> `static ngx_inline` on the signatures so a consumer using part of
the family stays warning-clean. The header is declared in
HTTP_ZSTD_DEPS so incremental builds rebuild on edits.

No behaviour change; the 00-filter no-transform tests (TESTs from
#251/#274) pass unchanged.
